Distinguish between (no translation) and (empty string) cases
This allows to wipe out the translation for a specific item, leading to the placeholder string "no translation" in the UI and also removing the <target> tag entirely. After editing the field, an empty string is now also possible, this will lead to an empty <target> tag. To wipe out the translation (and also potentially an empty string), the user can use the BS key.

Note: unfortunately Xcode will treat both cases as (no translation), currently (8.1). The workaround is to use a single space for the translation string (this will _not_ be collapsed).

This resolves #9