XliffFile Filter Refactoring
instead of creating a brand new instance of XliffFile every time a new filter is applied, the XliffFile Class has now a filter property, so the existing instance can be re-used. This also solves the issue with the expanded-state not being preserved while using the filter and should also give a better performance while searching (because the XML parsing is done once and also because all objects are reused).