Handle \n in <trans-unit> id attribute well
Xcode generates id attributes like id="Label 3 Line1&#10;Line2&#10;Line3" when using NSLocalizedString() API calls having source string containing new lines.
Because XMLDocument() converts them to regular new lines, the saved XML file does not match the original one.
This leads to a bug when Xcode being unable to find the correct references and will break these translations.

This workaround uses a special escape char sequence to mark all occurrences of \n and convert them back after the XML export.
This is a kind-of hack, it is very unfortunate that none of the XMLNode.Options is able to fix that behavior properly..